Add no-op logging macros to the Sabre SDK
Logging does not currenlty compile into wasm. With these
macros a developer can write a log statment in a smart contract
that will be run normally if compiled into a transaction processor
and ignored in wasm code.

Before it was required to include the following above the log
statement:

`#[cfg(not(target_arch = "wasm32"))]`

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>